refactor: restructure extensions-related code (#664)
